Remove casting insanity
GetAppletVersion() returned an unsigned long, but that needed to be
stored in a CK_BYTE (aka unsigned char), which requires us to cast.

The actual data coming from the card is not an unsigned long, it is an
unsigned char.

That means we need to cast it from a byte to a long and back to a byte,
which makes no sense (and produces a warning on the second count).

Remove the unsigned long from the declarations, and turn it into an
unsigned char.